// features/empreendimentos/types/empreendimentoForm.ts
export interface EmpreendimentoFormData {
  nome: string;

  cidade: string;

  bairro: string;

  estado: string;

  cep: string;

  endereco: string;

  latitude: string;

  longitude: string;

  construtora: string;

  incorporadora: string;

  tipo: string;

  status: string;

  entrega: string;

  registro: string;

  valorInicial: string;

  valorFinal: string;

  areaInicial: string;

  areaFinal: string;

  vgv: string;

  descricao: string;

  diferenciais: string;

  infraestrutura: string;

  lazer: string;

  publico: string;
}